+++ .gitignore
@@ -159,7 +159,7 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-*.pt
+*.keras
 config.ini
 *.avi
 /train/detect_desk/datasets
