--- src/detect_desk/yolov5_trash_data.yaml
@@ -1,8 +1,0 @@
-path: ../../datasets/yolov5_trash
-
-train: train/images
-val: valid/images
-test: test/images
-
-nc: 5
-names: ['-', 'cardboard', 'glass', 'metal', 'plastic']
