--- src/detect_desk/yolov8_trash_data.yaml
@@ -1,15 +1,0 @@
-path: ../../datasets/yolov8_trash
-
-train: ../train/images
-val: ../valid/images
-test: ../test/images
-
-nc: 42
-names: ['Aerosols', 'Aluminum can', 'Aluminum caps', 'Cardboard', 'Cellulose', 'Ceramic', 'Combined plastic', 'Container for household chemicals', 'Disposable tableware', 'Electronics', 'Foil', 'Furniture', 'Glass bottle', 'Iron utensils', 'Liquid', 'Metal shavings', 'Milk bottle', 'Organic', 'Paper bag', 'Paper cups', 'Paper shavings', 'Paper', 'Papier mache', 'Plastic bag', 'Plastic bottle', 'Plastic can', 'Plastic canister', 'Plastic caps', 'Plastic cup', 'Plastic shaker', 'Plastic shavings', 'Plastic toys', 'Postal packaging', 'Printing industry', 'Scrap metal', 'Stretch film', 'Tetra pack', 'Textile', 'Tin', 'Unknown plastic', 'Wood', 'Zip plastic bag']
-
-roboflow:
-  workspace: projectverba
-  project: yolo-waste-detection
-  version: 1
-  license: CC BY 4.0
-  url: https://universe.roboflow.com/projectverba/yolo-waste-detection/dataset/1
