--- src/detect_phone/src/haejo.ui
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Dialog</class>
- <widget class="QDialog" name="Dialog">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>1060</width>
-    <height>715</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Dialog</string>
-  </property>
-  <widget class="QPushButton" name="detect_snack">
-   <property name="geometry">
-    <rect>
-     <x>730</x>
-     <y>60</y>
-     <width>151</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>detect_snack</string>
-   </property>
-  </widget>
-  <widget class="QLabel" name="screen">
-   <property name="geometry">
-    <rect>
-     <x>110</x>
-     <y>180</y>
-     <width>771</width>
-     <height>461</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>TextLabel</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="detect_desk">
-   <property name="geometry">
-    <rect>
-     <x>570</x>
-     <y>60</y>
-     <width>151</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>detect_desk</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="detect_door">
-   <property name="geometry">
-    <rect>
-     <x>410</x>
-     <y>60</y>
-     <width>151</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>detect_door</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="detect_light">
-   <property name="geometry">
-    <rect>
-     <x>250</x>
-     <y>60</y>
-     <width>151</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>detect_light</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="detect_phone">
-   <property name="geometry">
-    <rect>
-     <x>90</x>
-     <y>60</y>
-     <width>151</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>detect_phone</string>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections/>
-</ui>
